Cosmos_embedding/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Configuration
python-dotenv==1.0.0

# Azure Cosmos DB
azure-cosmos==4.5.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
unstructured==0.11.8
docling==1.0.0
pypdf==4.0.1

# Embeddings and AI
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchain-core==0.1.52
tiktoken==0.5.2

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-magic==0.4.27
numpy==1.24.3

# Logging and Monitoring
loguru==0.7.2

